\name{bptwin}
\alias{bptwin}
\title{bptwin}
\description{
  bptwin
}
\usage{
bptwin(formula, data, id, zyg, twinnum, weight=NULL, ...)
}
\arguments{
  \item{formula}{}
  \item{data}{}
  \item{id}{}
  \item{zyg}{}
  \item{twinnum}{}
  \item{weight}{}
  \item{\dots}{}
}
\value{
  bptwin
}
\details{
  bptwin
}
\seealso{
  \code{twinlm}
}
\examples{
\dontrun{
bptwin(cancer~1,data=prt0,id="tvparnrs",zyg="zygbin",twinnum="tv",control=list(trace=1))
}
}
\author{Klaus K. Holst}
\keyword{regression}

